requirements.txt
flask>=3.0.0
pyyaml>=6.0
pytest>=7.0.0
pytest-flask>=1.2.0
